Maybe I should compile first, commit later?
git-svn-id: http://qeforge.qe-forge.org/svn/q-e/trunk/espresso@13607 c92efa57-630b-4861-b058-cf58834340f0

+++ TDDFPT/src/lr_eels_main.f90
@@ -46,7 +46,7 @@ PROGRAM lr_eels_main
   INTEGER             :: ip, na, pol_index, ibnd
   INTEGER             :: iter_restart, iteration
   LOGICAL             :: rflag
-  INTEGER(kind=c_int) :: kilobytes
+  INTEGER             :: kilobytes
   LOGICAL, EXTERNAL   :: test_restart
   !
   pol_index = 1
